% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22009.R
\name{dcor}
\alias{dcor}
\title{Computing brownian distance correlation using R}
\usage{
dcor(x, y)
}
\arguments{
\item{x}{a group of samples}

\item{y}{a group of samples}
}
\value{
the brownian distance correlation of the two groups of samples
}
\description{
Computing brownian distance correlation of the input samples using Rcpp
}
\examples{
\dontrun{
data(Eckerle4)
attach(Eckerle4)
dcor <- dcor(Eckerle4$wavelength, Eckerle4$transmitance)
}
}
