% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22009.R
\name{dcovR}
\alias{dcovR}
\title{Computing brownian distance covariance using R.}
\usage{
dcovR(x, y)
}
\arguments{
\item{x}{a group of samples}

\item{y}{a group of samples}
}
\value{
the brownian distance covariance of the two groups of samples
}
\description{
Computing brownian distance covariance of the input samples using Rcpp.
}
\examples{
\dontrun{
data(Eckerle4)
attach(Eckerle4)
dcov <- dcovR(Eckerle4$wavelength, Eckerle4$transmitance)
}
}
