% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22009.R
\name{Eckerle4}
\alias{Eckerle4}
\title{A illustrate dataset}
\description{
A dataset used to illustrate the brownian distance covariance of \code{dcovR} and \code{dcovC}.
}
\examples{
\dontrun{
data(Eckerle4)
attach(Eckerle4)
plot(Eckerle4)
}
}
