% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22009.R
\name{p_value}
\alias{p_value}
\title{Computing the p-value using R}
\usage{
p_value(x, y, cov_type, repl, Method = NULL)
}
\arguments{
\item{x}{a group of samples}

\item{y}{a group of samples}

\item{cov_type}{the tpye of covariance, cov() or dcov()}

\item{repl}{the times of replication}

\item{Method}{the tpye of covariance in cov(), pearson or spearman}
}
\value{
the p-value
}
\description{
Computing the p-value of the brownian distance covariance, pearson covariance and spearman covariance using R
}
\examples{
\dontrun{
data(Eckerle4)
attach(Eckerle4)
pvalue <- p_value(Eckerle4$wavelength, Eckerle4$transmitance, cov, 1000, 'pearson')
}
}
