\name{StatComp22009-package}
\alias{StatComp22009-package}
\alias{StatComp22009}
\docType{package}
\title{
  This package provides several functions for computing brownian distance covariance and P-value.
}
\description{
  R and Rcpp versions of brownian distance covariance, R version of P-value of different type of covariance.
}

\author{
Weijun Liang

Maintainer: Weijun Liang <liangweijun@mail.ustc.edu.cn>
}

\keyword{ package }
